# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Article API Documentation",
        "version": "0.1"
    },
    "paths": {
        "/login": {
            "post": {
                "description": "Logs user into system",
                "operationId": "Logs user into system",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/create": {
            "post": {
                "description": "Creates a new user",
                "operationId": "create user",
                "requestBody": {
                    "description": "Pet to add to the store",
                    "required": true,
                    "content": {
                        "multipart/form-data": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/showarticles": {
            "get": {
                "summary": "get article list",
                "operationId": "App\\Http\\Controllers\\Usercontroller::showArticle",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid tag value"
                    }
                }
            }
        },
        "/follow":{
            "post":{
            "summary": "follow",
            "operationId": "App\\Http\\Controllers\\Usercontroller::follow",
            "responses":{
           "200": {
            "description": "successfully followed"
           },
           "400":{
            "description": "Invalid tag value"
           }
           }
           }
        },
         "/unfollow":{
            "post":{
            "summary": "unfollow",
            "operationId": "App\\Http\\Controllers\\Usercontroller::unfollow",
            "responses":{
            "200":{
            "description": "successfully unfollowed"
            },
            "400":{
            "description": "Invalid tag value"
            }
        }
    }
},
           "/comment":{
           "post":{
           "summary": "comment on article",
           "operationId": "App\\Http\\Controllers\\Usercontroller::comment",
           "responses":{
           "200":{
           "description": "successful operation"
           },
           "400":{
          "description": "Invalid tag value"
           }
        }
    }
},
          
           "/update":{
           "put":{
           "summary": "update article",
           "operationId": "App\\Http\\Controllers\\Usercontroller::update",
           "responses":{
           "200":{
           "description": "successfully updated"
           },
           "400":{
            "description": "Invalid tag value"
           }
        }
    }
},
          "/adminView":{
          "delete":{
          "summary": "admin get article list",
          "operationId": "App\\Http\\Controllers\\Usercontroller::adminView",
          "responses":{
          "200":{
          "description": "successful operation"
          },
          "400":{
          "description": "Invalid tag value"
          }
        }
    }
},
          "/adminDeleteArticle":{
          "delete":{
          "summary": "admin can delete article", 
          "operationId": "App\\Http\\Controllers\\Usercontroller::adminDeleteArticle",
          "responses":{
          "200":{
           "description": "successful operation"
          },
          "400":{
           "description":"Invalid tag value"
          }
        }
    }
},
           "/logout":{
            "delete":{
            "summary": "logout", 
            "operationId": "App\\Http\\Controllers\\Usercontroller::logout",
            "responses":{
            "200":{
            "description": "successful operation"
            },
            "400":{
            "description": "Invalid tag value"
            }
          }
        }
      }
   }
}